Se cambio el tipo de dato de las coordenadas de Integer a Double

+++ src/main/java/tech/calaverita/sfast_xpress/DTOs/solicitud/UbicacionDTO.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class UbicacionDTO {
-    private Integer lat;
-    private Integer lng;
+    private Double lat;
+    private Double lng;
 }
